+++ myokit/tests/ansic_fixed_form_pacing.py
@@ -12,7 +12,7 @@ import os
 
 import myokit
 
-from shared import DIR_TEST
+from myokit.tests import DIR_TEST
 
 # Location of C template
 SOURCE_FILE = 'ansic_fixed_form_pacing.c'