+++ myokit/tests/test_rhs_benchmarker.py
@@ -25,7 +25,7 @@ class RhsBenchmarkerTest(unittest.TestCase):
     Tests :class:`RhsBenchmarker`.
     """
     def test_simple(self):
-        """ Test basic functionality. """
+        # Test basic functionality.
 
         # Create test model
         m = myokit.Model('test')
@@ -60,7 +60,7 @@ class RhsBenchmarkerTest(unittest.TestCase):
         mean, std = b.mean_std(t)
 
     def test_bad_log(self):
-        """ Test error handling when unsuitable log is used. """
+        # Test error handling when unsuitable log is used.
 
         # Create test model
         m = myokit.Model('test')
@@ -98,7 +98,7 @@ class RhsBenchmarkerTest(unittest.TestCase):
             ValueError, 'same length', b.bench_full, d, 1)
 
     def test_creation(self):
-        """ Test Benchmarker creation. """
+        # Test Benchmarker creation.
         # Create test model
         m = myokit.Model('test')
         c = m.add_component('c')